+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "s3widget",
+  "name": "joist",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
