--- django/s3widget/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class S3WidgetConfig(AppConfig):
-    name = 's3widget'
