+++ example-client/babel.config.js
