+++ example-client/package.json
