+++ example-client/public/favicon.ico
