+++ example-client/src/components/JoistUpload.vue
