+++ example-client/src/main.ts
