+++ example-client/src/shims-vue.d.ts
