+++ example-client/tsconfig.json
