+++ example/blobs/templates/base.html
