+++ example/blobs/templates/blob/detail.html
