+++ example/blobs/templates/blob/index.html
