+++ example/blobs/views.py
