+++ example/manage.py
@@ -14,7 +14,7 @@ def main() -> None:
         )
         dotenv.read_dotenv()
 
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'joist.settings')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'blobs.settings')
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
