+++ example/mypy.ini
@@ -5,4 +5,4 @@ plugins =
     mypy_drf_plugin.main
 
 [mypy.plugins.django-stubs]
-django_settings_module = joist.settings
+django_settings_module = blobs.settings
