+++ joist/templates/s3fileinput/s3fileinput.html
