+++ mypy.ini
@@ -1,0 +1,8 @@
+[mypy]
+ignore_missing_imports = True
+plugins =
+    mypy_django_plugin.main,
+    mypy_drf_plugin.main
+
+[mypy.plugins.django-stubs]
+django_settings_module = tests.settings
