+++ tests/settings.py
@@ -1,0 +1,8 @@
+SECRET_KEY = 'test'
+AWS_STORAGE_BUCKET_NAME = ''
+AWS_REGION = ''
+AWS_ACCESS_KEY_ID = ''
+AWS_SECRET_ACCESS_KEY = ''
+UPLOAD_STS_ARN = ''
+S3_WIDGET_UPLOAD_DURATION = 3600
+AWS_STORAGE_BUCKET_NAME = ''
