+++ tox.ini
@@ -18,8 +18,8 @@ deps =
     flake8-quotes
     pep8-naming
 commands =
-    flake8 {posargs:s3widget joist}
-    black --check {posargs:s3widget joist}
+    flake8 {posargs:joist example}
+    black --check {posargs:joist example}
 
 [testenv:typing]
 basepython = python3
@@ -34,7 +34,8 @@ deps =
 setenv =
     PYTHONPATH = {env:PYTHONPATH:}{:}{envsitepackagesdir}{:}{toxinidir}
 commands =
-    bash -c "source ./.env && mypy {posargs:s3widget joist}"
+    mypy {posargs:joist}
+    bash -c "cd example && source .env && mypy blobs"
 
 [testenv:format]
 basepython = python3
@@ -50,7 +51,7 @@ max-line-length = 100
 show-source = True
 # https://github.com/PyCQA/flake8-import-order
 import-order-style = appnexus
-application-import-names = core, joist
+application-import-names = joist
 ignore =
     # closing bracket does not match indentation of opening bracket’s line
     E123
